dfu_backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.28.0
python-multipart>=0.0.9
pydantic>=2.10.0
pillow>=10.4.0
numpy>=2.1.0
opencv-python-headless>=4.10.0.84
torch>=2.5.0
torchvision>=0.20.0
